requirements.txt
Django==1.5.1
argparse==1.2.1
django-tastypie==0.9.15
mimeparse==0.1.3
python-dateutil==2.1
redis==2.7.5
requests==1.2.3
six==1.3.0
wsgiref==0.1.2
